+++ src/server/listener.cc
@@ -1,6 +1,6 @@
 /* listener.cc
    Mathieu Stefani, 12 August 2015
-   
+
 */
 
 #include <iostream>
@@ -8,7 +8,7 @@
 #include <cstring>
 
 #include <sys/socket.h>
-#include <unistd.h> 
+#include <unistd.h>
 #include <netinet/in.h>
 #include <netinet/tcp.h>
 #include <arpa/inet.h>
@@ -147,7 +147,7 @@ Listener::bind(const Address& address) {
 
     struct addrinfo hints;
     hints.ai_family = AF_INET;
-    hints.ai_socktype = SOCK_STREAM; 
+    hints.ai_socktype = SOCK_STREAM;
     hints.ai_flags = AI_PASSIVE;
     hints.ai_protocol = 0;
 
@@ -220,7 +220,7 @@ Listener::run() {
                 else {
                     if (event.flags.hasFlag(Polling::NotifyOn::Read)) {
                         auto fd = event.tag.value();
-                        if (fd == listen_fd)
+                        if ((ssize_t)fd == listen_fd)
                             handleNewConnection();
                     }
                 }
